requirements.txt
python-dotenv
requests
replicate